pyproject.toml
[project]
name = "medical-assistant-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "langchain-groq>=1.1.2",
    "langchain-pinecone>=0.2.13",
    "pypdf>=6.9.2",
    "sentence-transformers>=5.3.0",
    "python-dotenv>=1.2.2",
    "ipykernel>=7.2.0",
    "flask>=3.1.3",
    "ipywidgets>=8.1.8",
    "notebook>=7.5.5",
    "pinecone>=7.3.0",
]
